# Improve string comparison performance