--- a/src/Wpf.Ui/Controls/EventIdentifier.cs
+++ b/src/Wpf.Ui/Controls/EventIdentifier.cs
@@ -3,8 +3,6 @@
 // Copyright (C) Leszek Pomianowski and WPF UI Contributors.
 // All Rights Reserved.
 
-using Wpf.Ui.Extensions;
-
 namespace Wpf.Ui.Controls;
 
 /// <summary>