--- a/src/Wpf.Ui/Controls/GridView/GridViewRowPresenter.cs
+++ b/src/Wpf.Ui/Controls/GridView/GridViewRowPresenter.cs
@@ -3,9 +3,7 @@
 // Copyright (C) Leszek Pomianowski and WPF UI Contributors.
 // All Rights Reserved.
 
-using System.Reflection;
-using System.Windows.Controls;
-
+// ReSharper disable once CheckNamespace
 namespace Wpf.Ui.Controls;
 
 /// <summary>