--- a/src/Wpf.Ui/Controls/IconElement/IconSourceElement.cs
+++ b/src/Wpf.Ui/Controls/IconElement/IconSourceElement.cs
@@ -4,7 +4,6 @@
 // All Rights Reserved.
 
 using System.Windows.Markup;
-using Wpf.Ui.Converters;
 
 // ReSharper disable once CheckNamespace
 namespace Wpf.Ui.Controls;