--- a/src/Wpf.Ui/Controls/TextBox/TextBox.cs
+++ b/src/Wpf.Ui/Controls/TextBox/TextBox.cs
@@ -5,7 +5,6 @@
 
 using System.Diagnostics;
 using System.Windows.Controls;
-using Wpf.Ui.Converters;
 using Wpf.Ui.Input;
 
 // ReSharper disable once CheckNamespace