--- a/src/Wpf.Ui/UiApplication.cs
+++ b/src/Wpf.Ui/UiApplication.cs
@@ -132,7 +132,7 @@ public class UiApplication
         return application
             .Resources.MergedDictionaries.Where(e => e.Source is not null)
             .Any(e =>
-                e.Source.ToString().ToLower().Contains(Appearance.ApplicationThemeManager.LibraryNamespace)
+                e.Source.ToString().Contains(Appearance.ApplicationThemeManager.LibraryNamespace, StringComparison.OrdinalIgnoreCase)
             );
     }
 }