The current version of the comparison is extremely inefficient. As you know `ToLower` is a rather slow operation, which may reduce performance for comparing a large number of strings. The best way is to use overloading with `StringComparison` support. But it is not available in **Net framework**.

I added the code and description from .Net 9 to support the method on any target framework. To prevent **using** from being marked unused in **Net core** and accidentally deleted, I added global using for extensions in the UI package.

Additionally I removed the `Trim` call, for URI comparison, URI does not support spaces and such call is useless, every call I debugged, everything compares correctly, previous behaviour is preserved with performance increase.

Before:
```C#
"text".Trim().ToLower().Contains("Text")
```

Now:
```C#
"text".Contains("Text", StringComparison.OrdinalIgnoreCase)
```

Benchmark for string `pack://application:,,,/Wpf.Ui;component/Resources/Theme/Light.xaml`:

| Method                     | Mean     | Error    | StdDev   | Allocated |
|--------------------------- |---------:|---------:|---------:|----------:|
| ToLower                    | 37.22 ns | 2.117 ns | 3.103 ns |     160 B |
| OrdinalIgnoreCase          | 13.61 ns | 0.138 ns | 0.203 ns |         - |
| InvariantCultureIgnoreCase | 92.63 ns | 0.656 ns | 0.962 ns |         - |

`StringComparison.OrdinalIgnoreCase` 3 times faster without memory allocation.

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Update
- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes